fix(bal-4087): magic links not working when url leads to 404 page
