+++ apps/backoffice-v2/src/pages/Root/Root.loader.ts
@@ -1,9 +1,34 @@
 import { LoaderFunction, redirect } from 'react-router-dom';
+import { queryClient } from '@/lib/react-query/query-client';
+import { magicLinkSignIn } from '@/domains/auth/fetchers';
 
-export const rootLoader: LoaderFunction = ({ request }) => {
+export const rootLoader: LoaderFunction = async ({ request }) => {
   const url = new URL(request.url);
 
-  if (url.pathname.startsWith('/en')) return null;
+  const token = url.searchParams.get('token');
 
-  return redirect(`/en${url.pathname === '/' ? '' : url.pathname}${url.search}`);
+  if (token) {
+    try {
+      await queryClient.fetchQuery(['magic-link-auth', token], ({ queryKey }) =>
+        magicLinkSignIn({ token: queryKey[1]! }),
+      );
+    } catch (e) {
+      console.error('Error using magic link', e);
+      return redirect(`/en/auth/sign-in`);
+    }
+  }
+
+  if (url.pathname.startsWith('/en') && !token) return null;
+
+  const newUrl = new URL(request.url);
+
+  if (token) {
+    newUrl.searchParams.delete('token');
+  }
+
+  if (!newUrl.pathname.startsWith('/en')) {
+    url.pathname = `/en${url.pathname === '/' ? '' : url.pathname}`;
+  }
+
+  return redirect(newUrl.toString());
 };
